Moved XMPUtilMain to cli
