+++ src/test/java/net/sf/jabref/logic/xmp/XMPUtilTest.java
@@ -27,7 +27,7 @@ import java.util.TimeZone;
 import javax.xml.transform.TransformerException;
 
 import net.sf.jabref.Globals;
-import net.sf.jabref.XMPUtilMain;
+import net.sf.jabref.cli.XMPUtilMain;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.importer.fileformat.BibtexParser;
 import net.sf.jabref.logic.bibtex.BibEntryWriter;
